requirements.txt
smartapi-python
pyotp
pandas
streamlit
plotly
requests
websocket-client
logzero
